Use bibtex-end-of-entry to search for closing parenthesis.
The hack that this commit replaces (searching for the next entry and then moving
back until a closing parenthesis is found) fails when the .bib file contains
material that is not part of any entry. BibTeX allows this, so parsebib should
as well.

Note that bibtex-end-of-entry works by parsing the entire entry, so it's
probably not as efficient. A better solution would be to get rid of the function
parsebib-match-paren-forward and detect the end of an entry through
parsebib--find-bibtex-field.

See also #8 for discussion.