Use symbols as keys in parsed json data.
The JSON parsing functions in Emacs return the field names as symbols, not as
strings. This differs from the .bib-parser, which returns the keys (field names)
as strings.

We're not going to change this. `bibtex.el` uses strings for field names, so
it's easier if `parsebib` returns strings in this case. OTOH the JSON libraries
not only return symbols as keys, the encoding functions also expect them.

The code calling `parsebib` should know whether it's parsing a .bib or a .json
file and should handle the return data accordingly.