Rename functions dealing with prettifying TeX markup.
Use a consistent prefix 'parsebib--TeX' for all internal functions
involved. Note that public variables already have the prefix
'parsebib-TeX-'. The one exception is 'parsebib-clean-TeX-markup', which is
a public function without 'parsebib-TeX' as prefix. We do not change it,
though, because it is public, so as not to affect packages depending on parsebib.