Improve error messages.
This reintroduces the detail in the error messages that was removed in
commit bd0b38e5. The cause of the slow-down was not the `format` function,
but rather the fact that each error message calculated the line and column
number. Now, instead, the position of point is added to the error data,
which is used to add positional data when the error message is printed.